requirements.txt
streamlit>=1.24.0
openai>=1.2.0
requests>=2.28.1
python-dotenv>=0.21.0
